Move addressing tests into respective modules
